+++ apps/shared/storage_cartesian_function.h
@@ -16,7 +16,7 @@ public:
   {}
   bool displayDerivative() const;
   void setDisplayDerivative(bool display);
-  int derivativeNameWithArgument(char * buffer, size_t bufferSize, char arg);
+  size_t derivativeNameWithArgument(char * buffer, size_t bufferSize, char arg);
   double approximateDerivative(double x, Poincare::Context * context) const;
   double sumBetweenBounds(double start, double end, Poincare::Context * context) const override;
   Poincare::Expression::Coordinate2D nextMinimumFrom(double start, double step, double max, Poincare::Context * context) const;
